requirements.txt
python==3.8.10
django==3.2.5
selenium==3.141.0
lxml==4.6.3
beautifulsoup4==4.9.3
requests==2.25.1
python-dotenv==0.19.0